refactor: remove chain-id input from workflow and update dependencies for improved compatibility
