+++ safe-transaction/package.json
@@ -2,19 +2,16 @@
   "name": "safe-transaction-action",
   "version": "1.0.0",
   "description": "GitHub Action for proposing Safe transactions",
+  "type": "module",
   "main": "dist/index.js",
   "scripts": {
-    "build": "ncc build src/index.js -o dist",
-    "prepare": "npm run build"
+    "build": "tsc"
   },
   "dependencies": {
     "@actions/core": "^1.10.1",
-    "@safe-global/api-kit": "^1.0.0",
-    "@safe-global/protocol-kit": "^3.0.0",
-    "@safe-global/types-kit": "^1.0.0",
-    "ethers": "^6.11.0"
+    "@safe-global/sdk-starter-kit": "^1.0.0"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.38.0"
+    "typescript": "^5.0.0"
   }
 }
